+++ lib/yass/cli/helpers.rb
@@ -49,6 +49,7 @@ yass <command> [path/to/dir] [options]
           ).strip
           opts.on("--clean", "Remove unknown files from dist/ when bulding") { config.clean = true }
           opts.on("--dest=DIR", "Build to a different directory") { |dir| config.dest = find_path(dir, cwd: Dir.pwd) }
+          opts.on("--drafts", "Include unpublished files in build") { config.include_drafts = true }
           opts.on("--no-strip-index", "Disable the strip_index Liquid filter") { config.strip_index = false }
           opts.on("--debug", "Print stack traces") { config.debug = true }
           opts.on("-h", "--help", "Prints this help") { config.stdout.puts opts; exit }
@@ -63,6 +64,7 @@ yass <command> [path/to/dir] [options]
           templates: "templates",
           dest: "dist",
           clean: false,
+          include_drafts: false,
           strip_index: true,
           stdin: $stdin,
           stdout: $stdout,